feat: fix persistence logic

+++ src/main/java/at/qe/skeleton/internal/services/CreditCardService.java
@@ -24,7 +24,6 @@ public class CreditCardService {
 
   @Autowired private UserxService userxService;
 
-
   public void saveCreditCard(CreditCard creditCard) throws IllegalArgumentException {
     if (!validateDate(creditCard.getExpirationDate())) {
       throw new IllegalArgumentException("Invalid expiration date.");
@@ -37,10 +36,13 @@ public class CreditCardService {
   }
 
   public void deleteCreditCard(CreditCard creditCard) {
-    if(userxService.isPremium(creditCard.getUser())) {
-      throw new IllegalArgumentException("You are a premium user. You cannot delete your credit card.");
+    if (creditCard == null) {
+      throw new IllegalArgumentException("Credit card cannot be null.");
     }
-
+    if (userxService.isPremium(creditCard.getUser())) {
+      throw new IllegalArgumentException(
+          "You are a premium user. You cannot delete your credit card.");
+    }
     creditCardRepository.delete(creditCard);
   }
